// scolaire-front/src/app/services/models/administrative-agent-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AdministrativeAgentRequest {
  address: string;
  archive?: boolean;
  emailPerso: string;
  emailPro: string;
  firstName: string;
  id?: number;
  lastName: string;
  phoneNumber: string;
}